chore: apply forge fmt

+++ test/BYAC.t.sol
@@ -3,9 +3,8 @@ pragma solidity ^0.8.6;
 
 import "forge-std/Test.sol";
 
-
 interface IBYAC {
-    function mintApe(uint numberOfTokens) external payable;
+    function mintApe(uint256 numberOfTokens) external payable;
     function balanceOf(address owner) external view returns (uint256 balance);
 }
 
@@ -13,7 +12,7 @@ contract BYACTest is Test {
     string public RPC_URL = "https://eth-mainnet.g.alchemy.com/v2/Ea1NmKEyxzz-p3O_4WZFu5apdNXGtV0l";
     IBYAC iBYAC;
     address user1 = address(1);
-    
+
     function setUp() public {
         uint256 forkId = vm.createFork(RPC_URL, 12299047);
         vm.selectFork(forkId);
@@ -26,13 +25,12 @@ contract BYACTest is Test {
         vm.startPrank(user1);
         uint256 oldBalance = address(iBYAC).balance;
         assertEq(iBYAC.balanceOf(user1), 0);
-        for(int i = 0; i < 5; i++){
+        for (int256 i = 0; i < 5; i++) {
             iBYAC.mintApe{value: 1.6 ether}(20);
         }
         uint256 newBalance = address(iBYAC).balance;
         assertEq(oldBalance + 8 ether, newBalance);
         assertEq(iBYAC.balanceOf(user1), 100);
         vm.stopPrank();
-        
     }
 }
